+++ changelog.md
@@ -2,6 +2,7 @@
 
 ### [Latest]
 
+- Update atlas-ftag-tools and puma [#109](https://github.com/umami-hep/umami-preprocessing/pull/109)
 - Adding Sample Check in prepare Stage [#108](https://github.com/umami-hep/umami-preprocessing/pull/108)
 - Adding init functions and general clean-up [#107](https://github.com/umami-hep/umami-preprocessing/pull/107)
 - Adding pydoclint and Update Pre-Commit [#106](https://github.com/umami-hep/umami-preprocessing/pull/106)